requirements.txt
python-telegram-bot[job-queue]>=21.0,<23.0
apscheduler>=3.10,<4.0
python-dotenv>=1.0
aiohttp>=3.9
aiosqlite>=0.19
